Possibility to open previously saved files Restructured fole and command architecture
